kafka/producer_ws/requirements.txt
websocket-client
kafka-python==2.0.2